requirements.txt
pyserial~=3.5
thingspeak~=1.0.0
httplib2~=0.20.1
mysql-connector-python~=8.0.27
setuptools~=58.4.0